Реализованы комментарии. Добавлена Dto в запрос на лайк
