+++ src/main/java/io/flats/payload/CommentDtoPayload.java
@@ -6,11 +6,11 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class CommentDtoPayload {
-    protected long id_from;
     protected long id_to;
+    protected String commentText;
 
-    public CommentDtoPayload(long id_from, long id_to) {
-        this.id_from = id_from;
+    public CommentDtoPayload(long id_to, String commentText) {
         this.id_to = id_to;
+        this.commentText = commentText;
     }
 }
