+++ src/main/java/io/flats/payload/LikeDtoPayload.java
@@ -6,11 +6,9 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class LikeDtoPayload {
-    protected long id_from;
     protected long id_to;
 
-    public LikeDtoPayload(long id_from, long id_to) {
-        this.id_from = id_from;
+    public LikeDtoPayload(long id_to) {
         this.id_to = id_to;
     }
 }
